[FIX] Compensated foreground cutoff that happened after previous fix.

+++ src/frame_graph/frame_graph.hpp
@@ -220,12 +220,8 @@ namespace wr
 
 		inline void Resize(RenderSystem & render_system, std::uint32_t width, std::uint32_t height)
 		{
-			//double widthScalar = width / m_old_window_size_width;
-			//double heightScalar = height / m_old_window_size_height;
-
 			std::uint32_t widthScalar = (width * 100000u)  / m_old_window_size_width;
 			std::uint32_t heightScalar = (height * 100000u) / m_old_window_size_height;
-
 
 			for (decltype(m_num_tasks) i = 0; i < m_num_tasks; ++i)
 			{
@@ -242,8 +238,8 @@ namespace wr
 					{
 						if (m_rt_properties[i].value().m_width.Get().value() > 0u && m_rt_properties[i].value().m_width.Get().value() > 0u)
 						{
-							newWidth = uint32_t((m_rt_properties[i].value().m_width.Get().value() * widthScalar) / 100000);
-							newHeight = uint32_t((m_rt_properties[i].value().m_height.Get().value() * heightScalar) / 100000);
+							newWidth = (m_rt_properties[i].value().m_width.Get().value() * widthScalar) / 100000u;
+							newHeight = (m_rt_properties[i].value().m_height.Get().value() * heightScalar) / 100000u;
 						}
 					}
